requirements.txt
streamlit

gTTS

python-docx
PyPDF2

transformers

torch

